feat: add Nounish app generation functionality
Created `create-nounish-app` utility to generate a new Nounish app from a template. This utility takes the name of the new app as an input parameter, creates a new directory with the inputted name, and copies the template files into this new directory. Additional dependencies and changes are included in `tsconfig.json`, `pnpm-lock.yaml`, and `package.json`. A `.gitignore` file is also created to ignore `node_modules` and `dist`, as well as `.idea` and `.vscode`. The utility checks that the directory does not exist to avoid an unintentional overwrite.