chore: add `build` and `publish` scripts to `package.json` file
Included 'build', 'prepublishOnly', and 'publish' scripts to package.json to automate the building and publishing process of the Nounish app. These additions aim to streamline the process of creating and deploying the package by running TypeScript compile (tsc) before publishing, ensuring only the compiled and not the raw files are published.